add GUI component system

+++ src/math/rectangle.cpp
@@ -1,6 +1,6 @@
 #include "rectangle.hpp"
 
-bool IsInsideRectangle(const Rectangle& rect, Vec2<float> point)
+bool IsInsideRectangle(const Rectangle& rect, Vec2<uint32_t> point)
 {
     return rect.x0 <= point.x && point.x <= rect.x0 + rect.w &&
            rect.y0 <= point.y && point.y <= rect.y0 + rect.h;
